Abstract single selection logic to a base class
